{{cookiecutter.project_slug}}/django/requirements/base.txt
# -------------------------------------
# BASE REQUIREMENTS
# -------------------------------------

# Django Reqs
# =====================================

psycopg2-binary==2.7.4
mimeparse==0.1.3
pytz==2018.3
Django==1.11.11

# Django Apps
# ======================================

django-robots==3.0
django-environ==0.4.4
django-extensions==2.0.6
django-storages==1.6.5
django-model-utils==3.1.1
djangorestframework==3.7.7
django-ipware==2.0.1
django-crispy-forms==1.7.2
django-parler==1.9.2

# DjangoCMS Reqs
# =====================================

six==1.11.0
Pillow==5.0.0
html5lib>=0.999999,<0.99999999
django-filer==1.3.0
django-cms==3.5.2

# DjangoCMS plugins
# =====================================

cmsplugin-filer==1.1.3
djangocms-text-ckeditor==3.5.3
djangocms-history==0.5.3

# DjangoCMS Extra Reqs
# ====================================

django-taggit==0.22.2
django-taggit-autosuggest==0.3.2
django-taggit-templatetags==0.2.5
django-meta==1.4.1

# Project Reqs
# =====================================

pytest-django==3.1.2
pytest-pythonpath==0.7.2
boto==2.48.0
whitenoise==3.3.1
gunicorn==19.7.1
celery==4.1.0
requests==2.18.4
beautifulsoup4==4.6.0
geoip2==2.7.0
geopy==1.12.0
{% if cookiecutter.use_uploadcare.lower() == "y" %}
pyuploadcare==2.2.1
{% endif %}
